Add files via upload
designed two nodes each has a capability of controlling 4 devices but the number of the device can be increased very easily from the schematic. I divided the schematic into 3 parts. One for the node, one for the gateway and another for the button panel. I attached all the schematic herewith